+++ alkymi/utils.py
@@ -1,20 +1,18 @@
 import asyncio
+import concurrent.futures
 import subprocess
 import sys
 import threading
-from typing import List, TextIO, Optional, TypeVar, Callable
-
-from .logging import log
+from typing import List, TextIO, Optional, TypeVar, Callable, IO
 
 
 def call(args: List[str], echo_error_to_stream: Optional[TextIO] = sys.stderr,
          echo_output_to_stream: Optional[TextIO] = sys.stdout) -> subprocess.CompletedProcess:
     """
-    Utility command to run a system command and return the result (stdout and stderr will also be printed to the debug
-    log)
+    Utility command to run a system command and return the result
 
     :param args: The arguments representing the command to run, e.g. ["echo", "test"]
-    :param echo_error_to_stream: A stream to which to echo the call's stderr on a non-zero exit code
+    :param echo_error_to_stream: A stream to which to echo the call's stderr while the command is executing
     :param echo_output_to_stream: A stream to which to echo the call's stdout while the command is executing
     :return: The result of the execution as a subprocess.CompletedProcess instance
     """
@@ -22,8 +20,16 @@ def call(args: List[str], echo_error_to_stream: Optional[TextIO] = sys.stderr,
     if echo_output_to_stream is not None and getattr(echo_output_to_stream, "name", None) == sys.stdout.name:
         echo_output_to_stream = sys.stdout
 
+    # If running through a Lab CLI, sys.stderr may have been redirected
+    if echo_error_to_stream is not None and getattr(echo_error_to_stream, "name", None) == sys.stderr.name:
+        echo_error_to_stream = sys.stderr
+
+    # Use shorthands - note that the full check is still used in some places below to satisfy mypy
+    live_stdout = echo_output_to_stream is not None
+    live_stderr = echo_error_to_stream is not None
+
     # Buffer one line at a time if echoing live, otherwise just use the default
-    buffer_size = 1 if echo_output_to_stream is not None else -1
+    buffer_size = 1 if (live_stdout or live_stderr) else -1
     proc = subprocess.Popen(args, universal_newlines=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE,
                             bufsize=buffer_size)
 
@@ -31,38 +37,64 @@ def call(args: List[str], echo_error_to_stream: Optional[TextIO] = sys.stderr,
     assert proc.stdout is not None
     assert proc.stderr is not None
 
+    # Variables that will hold the read stdout and stderr
     stdout: str = ""
-    if echo_output_to_stream is not None:
-        # Print each line to the stream as it arrives
-        while proc.poll() is None:
-            line = proc.stdout.readline()
-            echo_output_to_stream.write(line)
-            stdout += line
+    stderr: str = ""
 
-        # Program has finished executing, check if any part of stdout still needs to be piped
-        line = proc.stdout.readline()
-        if line:
-            echo_output_to_stream.write(line)
-            stdout += line
+    # If reading both stdout and stderr, use threads to avoid blocking
+    if live_stdout and live_stderr:
+        assert echo_output_to_stream is not None
+        assert echo_error_to_stream is not None
+        with concurrent.futures.ThreadPoolExecutor(max_workers=2) as executor:
+            stdout_fut = executor.submit(_tee_pipe, proc, proc.stdout, echo_output_to_stream)
+            stderr_fut = executor.submit(_tee_pipe, proc, proc.stderr, echo_error_to_stream)
+            stdout = stdout_fut.result()
+            stderr = stderr_fut.result()
+    elif live_stdout:
+        assert echo_output_to_stream is not None
+        stdout = _tee_pipe(proc, proc.stdout, echo_output_to_stream)
+    elif live_stderr:
+        assert echo_error_to_stream is not None
+        stderr = _tee_pipe(proc, proc.stderr, echo_error_to_stream)
     else:
-        # Otherwise, simply wait for the command to finish and then grab stdout
+        # Otherwise, simply wait for the command to finish and then grab stdout and/or stderr
         proc.wait()
         stdout = proc.stdout.read()
-
-    stderr = proc.stderr.read()
-
-    # Send to alkymi log
-    log.debug(stdout)
-    log.debug(stderr)
-
-    # Always forward error messages to stderr if needed
-    if echo_error_to_stream is not None and proc.returncode != 0:
-        echo_error_to_stream.write(stderr)
+        stderr = proc.stderr.read()
 
     # Raise an error on failure
     if proc.returncode != 0:
         raise subprocess.CalledProcessError(proc.returncode, proc.args, stdout, stderr)
     return subprocess.CompletedProcess(proc.args, proc.returncode, stdout, stderr)
+
+
+def _tee_pipe(proc: subprocess.Popen, input_stream: IO[str], output_stream: TextIO) -> str:
+    """
+    Reads the content of an input stream while the provided process is running, and echoes it to the provided output
+    stream in real time, while also collecting everything read into a string that is returned once the process has
+    finished executing and all output has been read
+
+    :param proc: The process that produces the output
+    :param input_stream: The stream to read from (should be `proc.stdout` or `proc.stderr`)
+    :param output_stream: The stream to write output to
+    :return: Everything that was read
+    """
+    content: str = ""
+
+    # Program is executing, echo lines as they come in
+    while proc.poll() is None:
+        line = input_stream.readline()
+        output_stream.write(line)
+        content += line
+
+    # Program has finished executing, check if any part of stdout or stderr still needs to be piped
+    line = " "
+    while line:
+        line = input_stream.readline()
+        output_stream.write(line)
+        content += line
+
+    return content
 
 
 T = TypeVar('T')
@@ -77,10 +109,10 @@ def run_on_thread(func: Callable[..., T]) -> T:
 
     # Mutable object used to store result
     result: List[T] = []
-    ex = []
+    ex: List[Exception] = []
 
     def _call_and_set_result():
-        nonlocal result
+        nonlocal result, ex
         try:
             result.append(func())
         except Exception as e: